requirements.txt
asgiref==3.5.0
autopep8==1.6.0
boto3==1.21.38
botocore==1.24.38
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
decorator==5.1.1
Django==3.2.11
django-cors-headers==3.11.0
django-environ==0.8.1
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djongo==1.3.6
dnspython==2.2.1
drf-yasg==1.20.0
dynaconf==3.1.7
hvac==0.11.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==1.0.0
knox==0.1.14
loguru==0.6.0
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
Pillow==9.0.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pymongo==3.12.3
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
shortuuid==1.0.8
simplejson==3.17.6
six==1.16.0
sqlparse==0.2.4
toml==0.10.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
validators==0.18.2
win32-setctime==1.1.0
